pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "leo"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock (>=3.14.1,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

